Extract ITableObject, IRepository, and comb-guid generation into a Data base project
