+++ src/Data/Entities/ITableObject.cs
